Move hooks to header
